etl/requirements.txt
google-cloud-spanner
google-cloud-storage
vertexai
google-cloud-aiplatform
pandas

